change size_t to unsigned

+++ include/fluxions_gte_image.hpp
@@ -43,9 +43,9 @@ namespace Fluxions
 	{
 	public:
 		std::vector<ColorType> pixels;
-		size_t imageWidth;
-		size_t imageHeight;
-		size_t imageDepth;
+		unsigned imageWidth;
+		unsigned imageHeight;
+		unsigned imageDepth;
 
 		using ImageType = TImage<ColorType>;
 
@@ -55,22 +55,22 @@ namespace Fluxions
 
 		constexpr void setBorderColor(const ColorType& color) { borderColor = color; }
 		constexpr const ColorType& getBorderColor() const { return borderColor; }
-		constexpr size_t width() const { return imageWidth; }
-		constexpr size_t height() const { return imageHeight; }
-		constexpr size_t depth() const { return imageDepth; }
-		constexpr size_t addr(size_t x, size_t y, size_t z = 0) {
+		constexpr unsigned width() const { return imageWidth; }
+		constexpr unsigned height() const { return imageHeight; }
+		constexpr unsigned depth() const { return imageDepth; }
+		constexpr unsigned addr(unsigned x, unsigned y, unsigned z = 0) {
 			if (x >= imageWidth || y >= imageHeight || z >= imageDepth)
 				return 0;
 			return z * zstride + y * imageWidth + x;
 		}
 
-		constexpr ColorType* getPixels(size_t index) {
+		constexpr ColorType* getPixels(unsigned index) {
 			if (index >= imageDepth)
 				return NULL;
 			return &(pixels[index * zstride]);
 		}
 
-		void resize(size_t width, size_t height, size_t depth = 1);
+		void resize(unsigned width, unsigned height, unsigned depth = 1);
 		void reset() { resize(0, 0, 0); }
 		void clear(const ColorType& clearcolor);
 		bool empty() const noexcept { return pixels.empty(); }
@@ -80,15 +80,15 @@ namespace Fluxions
 		TImage<ColorType>& ToneMap(float exposure);
 		TImage<ColorType>& ReverseToneMap(float exposure);
 		void scaleColors(float x);
-		TImage<ColorType> ScaleImage(size_t newWidth, size_t newHeight, bool bilinear = false);
+		TImage<ColorType> ScaleImage(unsigned newWidth, unsigned newHeight, bool bilinear = false);
 
-		constexpr void setPixel(size_t x, size_t y, ColorType color) noexcept {
+		constexpr void setPixel(unsigned x, unsigned y, ColorType color) noexcept {
 			if (x >= imageWidth || y >= imageHeight)
 				return;
 			pixels[y * imageWidth + x] = color;
 		}
 
-		constexpr ColorType getPixel(size_t x, size_t y) const {
+		constexpr ColorType getPixel(unsigned x, unsigned y) const {
 			if (x >= imageWidth || y >= imageHeight)
 				return borderColor;
 
@@ -96,22 +96,22 @@ namespace Fluxions
 		}
 
 		// Same as setPixel(X, y, color) but without image bounds checking: unsafe!
-		constexpr void setPixelUnsafe(size_t x, size_t y, ColorType color) {
+		constexpr void setPixelUnsafe(unsigned x, unsigned y, ColorType color) {
 			pixels[y * imageWidth + x] = color;
 		}
 
 		// Same as getPixel(X, y) but without image bounds checking: unsafe!
-		constexpr ColorType getPixelUnsafe(size_t x, size_t y) const {
+		constexpr ColorType getPixelUnsafe(unsigned x, unsigned y) const {
 			return pixels[y * imageWidth + x];
 		}
 
-		constexpr void setPixel(size_t x, size_t y, size_t z, const ColorType& color) {
+		constexpr void setPixel(unsigned x, unsigned y, unsigned z, const ColorType& color) {
 			if (x >= imageWidth || y >= imageHeight || z >= imageDepth)
 				return;
 			pixels[z * zstride + y * imageWidth + x] = color;
 		}
 
-		constexpr ColorType getPixel(size_t x, size_t y, size_t z) const {
+		constexpr ColorType getPixel(unsigned x, unsigned y, unsigned z) const {
 			if (x >= imageWidth || y >= imageHeight || z >= imageDepth)
 				return borderColor;
 
@@ -144,24 +144,24 @@ namespace Fluxions
 
 			float s = 0.0f;
 			float t = 0.0f;
-			size_t iz = 0;
+			unsigned iz = 0;
 			MakeFaceSTFromCubeVector(x, y, z, &s, &t, &iz);
-			size_t ix = (int)(s * imageWidth);
-			size_t iy = (int)((1.0f - t) * imageHeight);
+			unsigned ix = (int)(s * imageWidth);
+			unsigned iy = (int)((1.0f - t) * imageHeight);
 			return getPixel(ix, iy, iz);
 		}
 
 		// Same as setPixel(X, y, color) but without image bounds checking: unsafe!
-		constexpr void setPixelUnsafe(size_t x, size_t y, size_t z, const ColorType& color) {
+		constexpr void setPixelUnsafe(unsigned x, unsigned y, unsigned z, const ColorType& color) {
 			pixels[z * zstride + y * imageWidth + x] = color;
 		}
 
 		// Same as getPixel(X, y) but without image bounds checking: unsafe!
-		constexpr ColorType getPixelUnsafe(size_t x, size_t y, size_t z) const {
+		constexpr ColorType getPixelUnsafe(unsigned x, unsigned y, unsigned z) const {
 			return pixels[z * zstride + y * imageWidth + x];
 		}
 
-		constexpr const void* getImageData(size_t z) const {
+		constexpr const void* getImageData(unsigned z) const {
 			return &pixels[z * zstride];
 		}
 
@@ -171,10 +171,10 @@ namespace Fluxions
 		double getAverage() const;
 		int getNumPixels() const { return (int)pixels.size(); }
 
-		void savePPMRaw(const std::string& filename, size_t z = 0) const;
-		void savePPM(const std::string& filename, size_t z = 0, bool flipy = false) const;
-		void savePPMi(const std::string& filename, float scale, int minValue, int maxValue, size_t z = 0, bool flipy = false) const;
-		void savePPMHDRI(const std::string& filename, size_t z = 0) const;
+		void savePPMRaw(const std::string& filename, unsigned z = 0) const;
+		void savePPM(const std::string& filename, unsigned z = 0, bool flipy = false) const;
+		void savePPMi(const std::string& filename, float scale, int minValue, int maxValue, unsigned z = 0, bool flipy = false) const;
+		void savePPMHDRI(const std::string& filename, unsigned z = 0) const;
 		void loadPPM(const std::string& filename);
 		void loadEXR(const std::string& path);
 		void saveEXR(const std::string& path) const;
@@ -189,12 +189,12 @@ namespace Fluxions
 		bool convertRectToCubeMap(ImageType& dst) const;
 		bool convertCubeMapToRect(ImageType& dst) const;
 
-		void setImageData(unsigned int format, unsigned int type, size_t width, size_t height, size_t depth, void* _pixels);
+		void setImageData(unsigned int format, unsigned int type, unsigned width, unsigned height, unsigned depth, void* _pixels);
 
 	private:
-		void _setImageData(unsigned int fromFormat, unsigned int fromType, unsigned int toFormat, unsigned int toType, size_t width, size_t height, size_t depth, void* _pixels);
+		void _setImageData(unsigned int fromFormat, unsigned int fromType, unsigned int toFormat, unsigned int toType, unsigned width, unsigned height, unsigned depth, void* _pixels);
 
-		size_t zstride;
+		unsigned zstride;
 		ColorType borderColor;
 		int minColor;
 		int maxColor;
